\name{coef.cdtmbnma}
\alias{coef.cdtmbnma}
\title{Extract Component Coefficients}
\usage{
\method{coef}{cdtmbnma}(object, ...)
}
\arguments{
\item{object}{A fitted \code{cdtmbnma} object.}
\item{...}{Unused.}
}
\value{A data frame with posterior summaries for component Emax and ED50 values.}
\description{Extracts component dose-response coefficient summaries.}
